# add support for bracketed IPv6 addresses in the IRI authority component