--- a/xurls.go
+++ b/xurls.go
@@ -161,10 +161,10 @@ func relaxedExp() string {
 	tlds := `(?:(?i)` + punycode + `|` + anyOf(append(asciiTLDs, PseudoTLDs...)...) + `\b|` + anyOf(unicodeTLDs...) + `)`
 	domain := subdomain + tlds
 
-	hostName := `(?:` + domain + `|` + ipAddrMinusEmpty + `)`
+	hostName := `(?:` + domain + `|\[` + ipv6Addr + `\]|\b` + ipv4Addr + `\b)`
 	webURL := hostName + port + `(?:/` + pathCont + `|/)?`
 	email := `[a-zA-Z0-9._%\-+]+@` + domain
-	return strictExp() + `|` + webURL + `|` + email
+	return strictExp() + `|` + webURL + `|` + email + `|` + ipv6AddrMinusEmpty
 }
 
 // Strict produces a regexp that matches any URL with a scheme in either the