Ref #62

includes changes from #63

`benchstat` vs. master@68fd825f4d226a5b043f697ed9c2c29011c4c9ac

    name                         old time/op    new time/op    delta
    Strict_none-4                  45.7µs ±27%    41.8µs ± 4%     ~     (p=0.315 n=10+8)
    Strict_many-4                  36.0µs ± 4%    36.8µs ± 6%     ~     (p=0.156 n=9+10)
    Relaxed_none-4                 60.8µs ±19%    68.2µs ±17%  +12.19%  (p=0.011 n=10+10)
    Relaxed_many-4                 50.1µs ± 7%    56.8µs ±18%  +13.37%  (p=0.000 n=9+10)
    StrictMatchingScheme_none-4     775ns ±39%     743ns ±24%     ~     (p=0.739 n=10+10)
    StrictMatchingScheme_many-4    1.77µs ± 9%    1.95µs ±17%     ~     (p=0.053 n=9+10)

    name                         old speed      new speed      delta
    Strict_none-4                1.73MB/s ±22%  1.87MB/s ± 4%     ~     (p=0.303 n=10+8)
    Strict_many-4                2.69MB/s ± 4%  2.64MB/s ± 6%     ~     (p=0.138 n=9+10)
    Relaxed_none-4               1.29MB/s ±16%  1.16MB/s ±15%  -10.60%  (p=0.011 n=10+10)
    Relaxed_many-4               1.94MB/s ± 6%  1.72MB/s ±16%  -11.34%  (p=0.000 n=9+10)
    StrictMatchingScheme_none-4   103MB/s ±30%   106MB/s ±20%     ~     (p=0.739 n=10+10)
    StrictMatchingScheme_many-4  53.9MB/s ±19%  50.4MB/s ±18%     ~     (p=0.123 n=10+10)

    name                         old alloc/op   new alloc/op   delta
    Strict_none-4                   0.00B          0.00B          ~     (all equal)
    Strict_many-4                    230B ± 7%      233B ± 4%     ~     (p=0.790 n=10+10)
    Relaxed_none-4                 8.80B ±241%     0.00B          ~     (p=0.173 n=10+9)
    Relaxed_many-4                   240B ± 0%      240B ± 0%     ~     (all equal)
    StrictMatchingScheme_none-4     0.00B          0.00B          ~     (all equal)
    StrictMatchingScheme_many-4      192B ± 0%      192B ± 0%     ~     (all equal)

    name                         old allocs/op  new allocs/op  delta
    Strict_none-4                    0.00           0.00          ~     (all equal)
    Strict_many-4                    5.00 ± 0%      5.00 ± 0%     ~     (all equal)
    Relaxed_none-4                   0.00           0.00          ~     (all equal)
    Relaxed_many-4                   6.00 ± 0%      6.00 ± 0%     ~     (all equal)
    StrictMatchingScheme_none-4      0.00           0.00          ~     (all equal)
    StrictMatchingScheme_many-4      3.00 ± 0%      3.00 ± 0%     ~     (all equal)